requirements.txt
asn1crypto
certifi
cffi
charset-normalizer
cloudpickle
cryptography
filelock
idna
oscrypto
pycparser
pycryptodomex
PyJWT
pyOpenSSL
pytest
pytz
requests
snowflake-connector-python
snowflake-snowpark-python
typing-extensions
urllib3
